emulator: move to emulatorold for now
