+++ src/Snowflake.Framework.Primitives/Emulator/IEmulatorProvision.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Text;
+
+namespace Snowflake.Emulator
+{
+    public interface IEmulatorProvision
+    {
+        Guid ProvisionGuid { get; }
+        DirectoryInfo ProvisionRoot { get; }
+    }
+}
