+++ src/Snowflake.Framework.Primitives/EmulatorOld/IEmulatedGameProvision.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text;
 
-namespace Snowflake.Emulator
+namespace Snowflake.EmulatorOld
 {
     public interface IEmulatedGameProvision
     {
