+++ src/Snowflake.Framework/EmulatorOld/EmulatorAdapter.cs
@@ -14,7 +14,7 @@ using Snowflake.Records.File;
 using Snowflake.Records.Game;
 using Snowflake.Services;
 
-namespace Snowflake.Emulator
+namespace Snowflake.EmulatorOld
 {
     public abstract class EmulatorAdapter : ProvisionedPlugin, IEmulatorAdapter
     {
