+++ src/Snowflake.Plugin.Emulators.RetroArch/Adapters/RetroArchCommonAdapter.cs
@@ -6,7 +6,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Snowflake.Configuration;
 using Snowflake.Configuration.Input;
-using Snowflake.Emulator;
+using Snowflake.EmulatorOld;
 using Snowflake.Extensibility;
 using Snowflake.Extensibility.Provisioning;
 using Snowflake.Input;
