+++ src/Snowflake.Plugin.Emulators.RetroArch/Adapters/bsnes/BsnesInstance.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Snowflake.Configuration;
-using Snowflake.Emulator;
+using Snowflake.EmulatorOld;
 using Snowflake.Platform;
 using Snowflake.Plugin.Emulators.RetroArch.Adapters.bsnes.Configuration;
 using Snowflake.Plugin.Emulators.RetroArch.Adapters.bsnes.Selections;
